Fixed all the 3 tests in GeoWaveBasicHBaseIT. All of them are passing now.

+++ extensions/datastores/hbase/src/main/java/mil/nga/giat/geowave/datastore/hbase/operations/BasicHBaseOperations.java
@@ -137,10 +137,12 @@ public class BasicHBaseOperations
 				tableNames,
 				tableNamesArr);
 		for (final TableName tableName : tableNames) {
-			conn.getAdmin().disableTable(
-					tableName);
-			conn.getAdmin().deleteTable(
-					tableName);
+			if(conn.getAdmin().isTableAvailable(tableName)){
+				conn.getAdmin().disableTable(
+						tableName);
+				conn.getAdmin().deleteTable(
+						tableName);
+			}
 		}
 	}
